requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.1.2
Django==1.11.2
django-floppyforms==1.7.0
django-taggit==0.22.1
django-taggit-serializer==0.1.5
djangorestframework==3.6.3
enum34==1.1.6
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.2
Pygments==2.2.0
pytz==2017.2
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
